requirements.txt
asgiref==3.5.2
asttokens==2.0.8
attrs==22.1.0
autopep8==1.7.0
backcall==0.2.0
boto3==1.24.76
botocore==1.27.76
certifi==2022.9.14
charset-normalizer==2.1.1
debugpy==1.6.3
decorator==5.1.1
Django==4.1.1
entrypoints==0.4
executing==1.0.0
flake8==5.0.4
idna==3.4
iniconfig==1.1.1
ipykernel==6.15.3
ipython==8.5.0
jedi==0.18.1
jmespath==1.0.1
jupyter-core==4.11.1
jupyter_client==7.3.5
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy==0.971
mypy-extensions==0.4.3
nest-asyncio==1.5.5
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pluggy==1.0.0
prompt-toolkit==3.0.31
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycodestyle==2.9.1
pyflakes==2.5.0
Pygments==2.13.0
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.3
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
PyYAML==6.0
pyzmq==23.2.1
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.5.0
toml==0.10.2
tomli==2.0.1
tornado==6.2
traitlets==5.4.0
twilio==7.14.0
types-PyYAML==6.0.11
types-requests==2.28.10
types-setuptools==65.3.0
types-urllib3==1.26.24
typing_extensions==4.3.0
urllib3==1.26.12
wcwidth==0.2.5
